Fix race condition
We forgot to clear queue observers and SM observers at StateMachine.trans_now(),
so we refactor the code because we do cancel_state_tasks() we meant to cancel
all event generators tied to the state.

This bug caused an exception when the routine that called trans_now() cancelled a
global timer. Before the state transition was completed by the async schedule_trans(),
an MQTT event handler tried to cancel the same timer, when a MQTT message arrived
at the right time to exercise this race condition.

(We raise an exception when a timer is cancelled twice to catch this kind of problem.)